backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.1
pytesseract==0.3.10
numpy==1.26.4
opencv-python==4.9.0.80 